I added the toggle-able item in the hud also with a better list for what the hud draws. Q is to toggle the item from potion to throwable (like in the GDD)

+++ MansionMashers/MansionMashers/SourceFiles/Sprite.c
@@ -68,7 +68,7 @@ AEGfxVertexList* createMesh(float width, float height, float offsetX, float offs
 	float halfWidth = width / 2;
 	float halfHeight = height / 2;
 
-	float len = sqrt(halfWidth*halfWidth + halfHeight*halfHeight);
+	double len = sqrt(halfWidth*halfWidth + halfHeight*halfHeight);
 
 	// Informing the library that we're about to start adding triangles
 	AEGfxMeshStart();
